feat(aci): Send project slug with test fire action request
The test notification endpoint accepts an optional projectSlug to
control which project is shown in the notification. This picks a
project from the connected monitors that the user has access to,
so the test notification displays a relevant project instead of an
arbitrary one.

Also extracts ActionFilterBlock into its own file for clarity.
