+++ static/app/views/automations/components/automationBuilder.tsx
@@ -2,36 +2,27 @@ import {useEffect, useMemo} from 'react';
 import styled from '@emotion/styled';
 
 import {Alert} from '@sentry/scraps/alert';
-import {Button} from '@sentry/scraps/button';
 import {Container, Flex} from '@sentry/scraps/layout';
 import {Select} from '@sentry/scraps/select';
 
 import {fetchOrgMembers} from 'sentry/actionCreators/members';
 import {ConditionBadge} from 'sentry/components/workflowEngine/ui/conditionBadge';
 import {PurpleTextButton} from 'sentry/components/workflowEngine/ui/purpleTextButton';
-import {IconAdd, IconDelete, IconMail} from 'sentry/icons';
+import {IconAdd} from 'sentry/icons';
 import {t, tct} from 'sentry/locale';
 import type {SelectValue} from 'sentry/types/core';
 import {
-  type DataConditionGroup,
   DataConditionGroupLogicType,
   DataConditionHandlerGroupType,
 } from 'sentry/types/workflowEngine/dataConditions';
-import type {RequestError} from 'sentry/utils/requestError/requestError';
 import {useApi} from 'sentry/utils/useApi';
 import {useOrganization} from 'sentry/utils/useOrganization';
-import {FILTER_MATCH_OPTIONS} from 'sentry/views/automations/components/actionFilters/constants';
-import {ActionNodeList} from 'sentry/views/automations/components/actionNodeList';
+import {ActionFilterBlock} from 'sentry/views/automations/components/actionFilterBlock';
 import {AutomationBuilderConflictContext} from 'sentry/views/automations/components/automationBuilderConflictContext';
 import {useAutomationBuilderContext} from 'sentry/views/automations/components/automationBuilderContext';
 import {useAutomationBuilderErrorContext} from 'sentry/views/automations/components/automationBuilderErrorContext';
-import {
-  stripActionFields,
-  validateActions,
-} from 'sentry/views/automations/components/automationFormData';
 import {DataConditionNodeList} from 'sentry/views/automations/components/dataConditionNodeList';
 import {TRIGGER_MATCH_OPTIONS} from 'sentry/views/automations/components/triggers/constants';
-import {useSendTestNotification} from 'sentry/views/automations/hooks';
 import {findConflictingConditions} from 'sentry/views/automations/hooks/utils';
 
 export function AutomationBuilder() {
@@ -129,131 +120,6 @@ export function AutomationBuilder() {
   );
 }
 
-interface ActionFilterBlockProps {
-  actionFilter: DataConditionGroup;
-}
-
-function ActionFilterBlock({actionFilter}: ActionFilterBlockProps) {
-  const {state, actions} = useAutomationBuilderContext();
-  const {setErrors} = useAutomationBuilderErrorContext();
-  const {mutate: sendTestNotification, isPending} = useSendTestNotification({
-    onError: (error: RequestError) => {
-      // Store test notification error in error context
-      setErrors(prev => ({
-        ...prev,
-        ...error?.responseJSON,
-      }));
-    },
-  });
-
-  const numActionFilters = state.actionFilters.length;
-
-  const handleSendTestNotification = () => {
-    const actionFilterActions = actionFilter.actions || [];
-
-    // Validate actions before sending test notification
-    const actionErrors = validateActions({actions: actionFilterActions});
-    setErrors(prev => ({...prev, ...actionErrors}));
-
-    // Only send test notification if there are no validation errors
-    if (Object.keys(actionErrors).length === 0) {
-      sendTestNotification(
-        actionFilterActions.map(action => {
-          return stripActionFields(action);
-        })
-      );
-    }
-  };
-
-  return (
-    <IfThenWrapper>
-      <Step>
-        <Flex direction="column" gap="md">
-          <StepLead data-test-id="action-filter-logic-type">
-            {tct('[if: If] [selector] of these filters match', {
-              if: <ConditionBadge />,
-              selector: (
-                <Container width="80px">
-                  <EmbeddedSelectField
-                    styles={{
-                      control: (provided: any) => ({
-                        ...provided,
-                        minHeight: '21px',
-                        height: '21px',
-                      }),
-                    }}
-                    isSearchable={false}
-                    isClearable={false}
-                    name={`actionFilters.${actionFilter.id}.logicType`}
-                    options={FILTER_MATCH_OPTIONS}
-                    size="xs"
-                    value={
-                      FILTER_MATCH_OPTIONS.find(
-                        choice =>
-                          choice.value === actionFilter.logicType ||
-                          choice.alias === actionFilter.logicType
-                      )?.value || actionFilter.logicType
-                    }
-                    onChange={(option: SelectValue<DataConditionGroupLogicType>) =>
-                      actions.updateIfLogicType(actionFilter.id, option.value)
-                    }
-                  />
-                </Container>
-              ),
-            })}
-          </StepLead>
-          {numActionFilters > 1 && (
-            <DeleteButton
-              aria-label={t('Delete If/Then Block')}
-              size="sm"
-              icon={<IconDelete />}
-              priority="transparent"
-              onClick={() => actions.removeIf(actionFilter.id)}
-              className="delete-condition-group"
-            />
-          )}
-          <DataConditionNodeList
-            handlerGroup={DataConditionHandlerGroupType.ACTION_FILTER}
-            label={t('Add filter')}
-            placeholder={t('Any event')}
-            groupId={actionFilter.id}
-            conditions={actionFilter?.conditions || []}
-            onAddRow={type => actions.addIfCondition(actionFilter.id, type)}
-            onDeleteRow={id => actions.removeIfCondition(actionFilter.id, id)}
-            updateCondition={(id, params) =>
-              actions.updateIfCondition(actionFilter.id, id, params)
-            }
-          />
-        </Flex>
-      </Step>
-      <Step>
-        <StepLead>
-          {tct('[then:Then] perform these actions', {
-            then: <ConditionBadge />,
-          })}
-        </StepLead>
-        <ActionNodeList
-          placeholder={t('Select an action')}
-          conditionGroupId={actionFilter.id}
-          actions={actionFilter?.actions || []}
-          onAddRow={handler => actions.addIfAction(actionFilter.id, handler)}
-          onDeleteRow={id => actions.removeIfAction(actionFilter.id, id)}
-          updateAction={(id, data) => actions.updateIfAction(actionFilter.id, id, data)}
-        />
-      </Step>
-      <span>
-        <Button
-          icon={<IconMail />}
-          onClick={handleSendTestNotification}
-          disabled={!actionFilter.actions?.length || isPending}
-        >
-          {t('Send Test Notification')}
-        </Button>
-      </span>
-    </IfThenWrapper>
-  );
-}
-
 const Step = styled(Flex)`
   flex-direction: column;
   gap: ${p => p.theme.space.sm};
@@ -269,31 +135,6 @@ const EmbeddedSelectField = styled(Select)`
   padding: 0;
   font-weight: ${p => p.theme.font.weight.sans.regular};
   text-transform: none;
-`;
-
-const IfThenWrapper = styled(Flex)`
-  position: relative;
-  flex-direction: column;
-  gap: ${p => p.theme.space.md};
-  border: 1px solid ${p => p.theme.tokens.border.primary};
-  border-radius: ${p => p.theme.radius.md};
-  padding: ${p => p.theme.space.lg};
-  margin-top: ${p => p.theme.space.md};
-
-  /* Only hide delete button when hover is supported */
-  @media (hover: hover) {
-    &:not(:hover):not(:focus-within) {
-      .delete-condition-group {
-        ${p => p.theme.visuallyHidden}
-      }
-    }
-  }
-`;
-
-const DeleteButton = styled(Button)`
-  position: absolute;
-  top: ${p => p.theme.space.sm};
-  right: ${p => p.theme.space.sm};
 `;
 
 const StyledAlert = styled(Alert)`
